deleted category and set posts cat ids to uncategorised id, fixes #12

+++ admin/delete-category.php
@@ -7,12 +7,17 @@ if(isset($_GET['id'])) {
 	
 	//for later
 	//update category _id of post that belong to this category to id of uncategorised category
+	$update_query = "UPDATE posts SET category_id=5 WHERE category_id=$id";
+	$update_result = mysqli_query($connection, $update_query);
 	
-	
-	//delete category
+	if(!mysqli_errno($connection)) {
+		//delete category
 	$query = "DELETE FROM categories WHERE id=$id LIMIT 1";
 	$result = mysqli_query($connection, $query);
 	$_SESSION['delete-category-success'] = "category deleted successfully";
+	}
+	
+	
 }
 
 header('location: ' . ROOT_URL . 'admin/manage-categories.php');
